Split ml_core.features into a package; rename TabularNwpFeatureEngineer → TabularFeatureEngineer
The old flat files (features.py at 835 lines, feature_engineer.py alongside it) were
confusingly named and hard to navigate. This pure rename/reorganise splits them into a
features/ package:

  feature_engineer.py          — FeatureEngineer(ABC) only
  tabular_feature_engineer.py  — TabularFeatureEngineer, _attach_nearest_nwp_cell,
                                  _engineer_features (was engineer_features), and the
                                  four tabular pipeline helpers
  _parsed_features.py          — typed feature descriptors + ParsedFeatures
  _nwp.py                      — upsampling, processing, and power/NWP join helpers
  _lags.py                     — power lag, weather lag, leaky-lag nullification
  __init__.py                  — re-exports FeatureEngineer and TabularFeatureEngineer

engineer_features is renamed _engineer_features (private) because the only production
path is through FeatureEngineer.engineer(); tests import it from the submodule directly.

No logic changes. All 138 tests pass.
